requirements.txt
aws-cdk-lib==2.21.1
constructs>=10.0.0,<11.0.0
pytest==6.2.5
